requirements.txt
streamlit
pandas
numpy
plotly
openpyxl
google-generativeai
streamlit-authenticator==0.3.3
bcrypt
PyYAML
